Improve error handling of SPF_dns_resolv_lookup
If we run out of memory, return TRY_AGAIN (TempError).
If there is a parse error, return NO_RECOVERY (PermError).

Returning partial results is not helpful because it may result in a
failure to match a sending host IP that was omitted from the parsed
DNS response and return a FAIL outcome.

There is already a check that we only process the RR types requested,
so errors won't be returned as a result of parsing other RR types.